(fix): fixed all f32 to real in math

+++ playgrounds/examples/mesh_generate.rs
@@ -13,8 +13,8 @@ fn draw_mesh(
             let p1 = display_data.vertices[indices[i] as usize] + position;
             let p2 = display_data.vertices[indices[(i + 1) % 3] as usize] + position;
             mode.draw_line_3D(
-                Vector3::new(p1.x(), p1.y(), p1.z()),
-                Vector3::new(p2.x(), p2.y(), p2.z()),
+                Vector3::new(p1.x() as f32, p1.y() as f32, p1.z() as f32),
+                Vector3::new(p2.x() as f32, p2.y() as f32, p2.z() as f32),
                 Color::new(
                     (display_data.color.x() * 255.0).clamp(0.0, 255.0) as u8,
                     (display_data.color.y() * 255.0).clamp(0.0, 255.0) as u8,
@@ -35,8 +35,8 @@ fn draw_polyline(
         let p1 = line_data.vertices[i] + position;
         let p2 = line_data.vertices[i + 1] + position;
         mode.draw_line_3D(
-            Vector3::new(p1.x(), p1.y(), p1.z()),
-            Vector3::new(p2.x(), p2.y(), p2.z()),
+            Vector3::new(p1.x() as f32, p1.y() as f32, p1.z() as f32),
+            Vector3::new(p2.x() as f32, p2.y() as f32, p2.z() as f32),
             Color::new(
                 (line_data.color.x() * 255.0).clamp(0.0, 255.0) as u8,
                 (line_data.color.y() * 255.0).clamp(0.0, 255.0) as u8,
@@ -107,7 +107,7 @@ fn main() {
         center: Vec3::zeros(),
         norm: Vec3::y_axis(),
         x_axis: Vec3::x_axis(),
-        range: (45f32.to_radians(), -90f32.to_radians()),
+        range: (45f64.to_radians(), -90f64.to_radians()),
     }, ::math::matrix::Vec4::from_xyzw(0.0, 1.0, 1.0, 1.0) , 100);
 
     let ellipse_arc = ConicArc {
@@ -117,7 +117,7 @@ fn main() {
             p: 2.0,
             position: Vec3::zeros(),
         }),
-        range: (45f32.to_radians(), -45f32.to_radians()),
+        range: (45f64.to_radians(), -45f64.to_radians()),
     };
     let ellipse_arc = conic_arc_to_display_data(&ellipse_arc, Vec4::from_xyzw(1.0, 1.0, 0.0, 1.0), 100);
 
@@ -129,11 +129,11 @@ fn main() {
             b: 1.0,
             position: Vec3::zeros(),
         }),
-        range: (45f32.to_radians(), -45f32.to_radians()),
+        range: (45f64.to_radians(), -45f64.to_radians()),
     };
     let ellipse_arc2 = conic_arc_to_display_data(&ellipse_arc2, Vec4::from_xyzw(0.0, 1.0, 1.0, 1.0), 100);
 
-    let ellipse_arc3 = ConicArcInPolar::new(2.0, 1.3, Vec3::zeros(), Vec3::x_axis(), Vec3::y_axis(), (45f32.to_radians(), -45f32.to_radians()));
+    let ellipse_arc3 = ConicArcInPolar::new(2.0, 1.3, Vec3::zeros(), Vec3::x_axis(), Vec3::y_axis(), (45f64.to_radians(), -45f64.to_radians()));
     let ellipse_arc3 = polar_conic_arc_to_display_data(&ellipse_arc3, Vec4::from_xyzw(0.5, 0.3, 1.0, 1.0), 100);
 
     while !rl.window_should_close() {
